requirements.txt
Django==2.2
psycopg2-binary==2.8.5
django-grappelli==2.6.5
django-filebrowser==3.5.7
sorl-thumbnail==12.2
django-simple-captcha==0.4.5
django-dajaxice==0.7
django-rosetta==0.7.6
django-autotranslate==0.2
django-chunked-upload==1.0.5
django-dbgettext==0.1.1
django-multiupload==0.1
django-localflavor==1.1
django-tinymce==1.5.3
django-crispy-forms==1.4.0
django-markdown==0.8.4
